Move auth key-generation-service imports
